// client/sipejam/store/types/auth.types.ts
export enum authActionTypes {
  SIGN_UP_START = "SIGN_UP_START",
  SIGN_UP_SUCCESS = "SIGN_UP_SUCCESS",
  SIGN_UP_FAILURE = "SIGN_UP_FAILURE",
  SIGN_IN_START = "SIGN_IN_START",
  SIGN_IN_SUCCESS = "SIGN_IN_SUCCESS",
  SIGN_IN_FAILURE = "SIGN_IN_FAILURE",
  OAUTH_START = "OAUTH_START",
  OAUTH_SUCCESS = "OAUTH_SUCCESS",
  OAUTH_FAILURE = "OAUTH_FAILURE",
}